super rough render skeleton, ray generation for camera, sample structure

+++ Source/GoblinRay.h
@@ -9,19 +9,19 @@ namespace Goblin {
     public:
         Ray();
         Ray(const Vector3& origin, const Vector3& dir, 
-            float start, float end = INFINITY, float depth = 0);
+            float start, float end = INFINITY, int depth = 0);
     public:
         Vector3 o;
         Vector3 d;
         float mint, maxt;
-        float depth;
+        int depth;
     };
     
     inline Ray::Ray(): o(Vector3::Zero), d(Vector3::Zero),
         mint(0), maxt(INFINITY), depth(0) {}
 
     inline Ray::Ray(const Vector3& origin, const Vector3& dir, 
-        float start, float end, float depth): o(origin), d(dir), 
+        float start, float end, int depth): o(origin), d(dir), 
         mint(start), maxt(end), depth(depth) {}
 
 }
